Add files via upload
Create a complete program that:
a. prompts a user to enter the name of an existing file.
b. If the file does not exist and or cannot be opened, the program should display an error
message and quit.
c. If the file exists and can be opened, prompt the user for a new file-name.
d. If the new file name already exists, print a message telling the user that the file already
exists and ask the user whether they want to APPEND
i. if the user enters ‘N’ for “NO”, print a message “Operation canceled by user”
and quit
ii. otherwise, open the file for APPEND and copy all of the contents from the
original file
e. If the file doesn’t already exist, create a new file with the name supplied by the user and
copy the contents of the original file into it.