Sort 3 variable c++
Write a complete working program that prompts a user to enter 3 integer values and then places those values in sorted-order.  The sort program should use a single ‘sort’ function that accepts two integer variable parameters (by reference) and tests to see if the first parameter is the smallest of the two parameters.  If so, the function does nothing, otherwise, the function will swap the values in the two variables.  The values can be sorted by comparing value 1 and 3, value 1 and 2, & finally, 2 and 3. */